final integration of longchain
