+++ server.js
@@ -10,15 +10,11 @@ import cloudinary from "cloudinary";
 import helmet from "helmet";
 import mongoSanitize from "express-mongo-sanitize";
 
-//llm langchain
-import { ChatGoogleGenerativeAI } from "@langchain/google-genai";
-import { ChatPromptTemplate } from "@langchain/core/prompts";
-import { StringOutputParser } from "@langchain/core/output_parsers";
-
 //importing routers
 import jobRouter from "./routes/jobRouter.js";
 import authRouter from "./routes/authRouter.js";
 import userRouter from "./routes/userRouter.js";
+import aiRouter from "./routes/aiRouter.js";
 
 // public folder
 import { dirname } from "path";
@@ -72,31 +68,33 @@ app.use(mongoSanitize());
 // app.get("/api/v1/test", (req, res) => {
 //   res.json({ msg: "test route" });
 // });
-app.post('/api/v1/ai/generate-text', async (req, res) => {
-  const DEFAULT_TEXT = `Tired of scattered job applications and missed opportunities? Take control of your career journey with our intuitive job tracker! Effortlessly organize your applications, track your progress, and manage important deadlines. Visualize your success, identify areas for improvement, and watch your career prospects grow. Let us help you land your dream job, one organized step at a time. Start tracking smarter today!`; // Your original text
 
-  try {
-    const model = new ChatGoogleGenerativeAI({
-      model: "gemini-2.0-flash",
-      apiKey: "AIzaSyBw3lrF_PSHxZP2FWPO731Ms1UbzTRxdnA",
-    });
+// app.post('/api/v1/ai/generate-text', async (req, res) => {
+//   const DEFAULT_TEXT = `Tired of scattered job applications and missed opportunities? Take control of your career journey with our intuitive job tracker! Effortlessly organize your applications, track your progress, and manage important deadlines. Visualize your success, identify areas for improvement, and watch your career prospects grow. Let us help you land your dream job, one organized step at a time. Start tracking smarter today!`; // Your original text
 
-    const prompt = ChatPromptTemplate.fromMessages([
-      ["system", "You are a helpful career advisor."],
-      ["human", `Generate a short, engaging paragraph (50-70 words) for a job tracking application landing page.
-        Focus on benefits like organization, career growth, and job management. Use professional but friendly tone.
-        Avoid markdown formatting.`]
-    ]);
+//   try {
+//     const model = new ChatGoogleGenerativeAI({
+//       model: "gemini-2.0-flash",
+//       apiKey: "AIzaSyBw3lrF_PSHxZP2FWPO731Ms1UbzTRxdnA",
+//     });
 
-    const chain = prompt.pipe(model).pipe(new StringOutputParser());
+//     const prompt = ChatPromptTemplate.fromMessages([
+//       ["system", "You are a helpful career advisor."],
+//       ["human", `Generate a short, engaging paragraph (50-70 words) for a job tracking application landing page.
+//         Focus on benefits like organization, career growth, and job management. Use professional but friendly tone.
+//         Avoid markdown formatting.`]
+//     ]);
 
-    const response = await chain.invoke({});
-    res.json({ text: response });
-  } catch (error) {
-    console.error('AI Error:', error);
-    res.json({ text: DEFAULT_TEXT });
-  }
-});
+//     const chain = prompt.pipe(model).pipe(new StringOutputParser());
+
+//     const response = await chain.invoke({});
+//     res.json({ text: response });
+//   } catch (error) {
+//     console.error('AI Error:', error);
+//     res.json({ text: DEFAULT_TEXT });
+//   }
+// });
+app.use("/api/v1/ai", aiRouter);
 
 app.use("/api/v1/auth", authRouter);
 app.use("/api/v1/jobs", authenticateUser, jobRouter); // we need to authenticate user before accessing jobs
